requirements.txt
pywin32>=306
tkinterdnd2>=0.3.0
pytest>=7.0
pyinstaller>=6.0
pillow>=10.0
